new table filed change

+++ ERPBLL/Agriculture/Interface/IRawMaterialIssueStockDetailsBusiness.cs
@@ -11,6 +11,6 @@ namespace ERPBLL.Agriculture.Interface
     {
 
         IEnumerable<RawMaterialIssueStockDetails> GetRawMaterialIssueStockDetailsById(long infoId, long orgId);
-        bool SaveIssuerawMaterialStockDetail(long OrganizationId, long RawMaterialId, int Quantity, string Unit, DateTime? IssueDate, DateTime? EntryDate, long? EntryUserId, DateTime? UpdateDate, long? UpdateUserId, string Status, long RawMaterialIssueStockId);
+        bool SaveIssuerawMaterialStockDetail(long OrganizationId, long RawMaterialId, double Quantity, string Unit, DateTime? IssueDate, DateTime? EntryDate, long? EntryUserId, DateTime? UpdateDate, long? UpdateUserId, string Status, long RawMaterialIssueStockId);
     }
 }
