+++ ERPBLL/Agriculture/RawMaterialIssueStockDetailsBusiness.cs
@@ -27,7 +27,7 @@ namespace ERPBLL.Agriculture
             return _rawMaterialIssueStockDetailsRepository.GetAll(i => i.OrganizationId == orgId && i.RawMaterialIssueStockId == infoId).ToList();
         }
 
-        public bool SaveIssuerawMaterialStockDetail(long OrganizationId, long RawMaterialId, int Quantity, string Unit, DateTime? IssueDate, DateTime? EntryDate, long? EntryUserId, DateTime? UpdateDate, long? UpdateUserId, string Status, long RawMaterialIssueStockId)
+        public bool SaveIssuerawMaterialStockDetail(long OrganizationId, long RawMaterialId, double Quantity, string Unit, DateTime? IssueDate, DateTime? EntryDate, long? EntryUserId, DateTime? UpdateDate, long? UpdateUserId, string Status, long RawMaterialIssueStockId)
         {
             List<RawMaterialIssueStockDetails> IssueRawMaterialStockDetail = new List<RawMaterialIssueStockDetails>();
