+++ ERPBO/Agriculture/DTOModels/FinishGoodProductionInfoDTO.cs
@@ -12,8 +12,8 @@ namespace ERPBO.Agriculture.DTOModels
         public string FinishGoodProductionBatch { get; set; }
         public string ReceipeBatchCode { get; set; }
         public long FinishGoodProductId { get; set; }
-        public int Quanity { get; set; }
-        public int TargetQuantity { get; set; }
+        public double Quanity { get; set; }
+        public double TargetQuantity { get; set; }
         public string Status { get; set; }
         public string Remarks { get; set; }
         public string flag { get; set; }
